Renamed files according to convention and other minor changes across the repository
