+++ lib/main.go
@@ -36,7 +36,6 @@ func main() {
 	case "tag":
 		log.Println("Starting Tag Client")
 		tag.Main()
-
 	default:
 		log.Println("No Main function for given arg")
 		return
