+++ pkg/protocol/tag/client.go
@@ -18,7 +18,6 @@ func Main() {
 	client.SetupClient("MainClient")
 	client.InsertSingleMockedReading("MainTest")
 	client.SendStateToServer()
-	// Test whatever here
 }
 
 func (client *Client) SetupClient(id string) {
@@ -51,7 +50,7 @@ func (client *Client) Insert(serialized []byte) error {
 func (client *Client) InsertSingleMockedReading(tagId string) {
 	reading := &pb.Reading{
 		TagId: tagId,
-		RpId:  "yy:yy:yy:yy:yy:yy",
+		RpId:  "rpId15",
 		Rssi:  69,
 		Ts: &timestamp.Timestamp{
 			Seconds: timestamppb.Now().Seconds,
