+++ pkg/protocol/utils/state_hander_test.go
@@ -1,6 +1,7 @@
 package utils
 
 import (
+	"fmt"
 	"log"
 	"reflect"
 	"sort"
@@ -14,7 +15,7 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
-func TestInnitStateHandler(t *testing.T) {
+func TestInitStateHandler(t *testing.T) {
 	sh := StateHandler{}
 	sh.InitStateHandler("666")
 
@@ -22,6 +23,7 @@ func TestInnitStateHandler(t *testing.T) {
 	assert.Equal(t, sh, shMock)
 
 }
+
 func TestGetReading(t *testing.T) {
 
 	mockReading := &pb.Reading{
@@ -179,6 +181,7 @@ func TestMultipleGetState(t *testing.T) {
 		log.Printf("Insert\n: Expected %v\n, Got %v\n", actualState, expectedMockState)
 	}
 }
+
 func TestUpdateReadingofSelf(t *testing.T) {
 	sh := StateHandler{}
 	sh.InitStateHandler("RandomUniqueId")
@@ -212,4 +215,409 @@ func generateMockReading(count int) []*pb.Reading {
 	}
 	return readings
 
+}
+
+func TestInsertSingleReading(t *testing.T) {
+
+	mockReading := &pb.Reading{
+		TagId: "MOCKTAG",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG2",
+		RpId:  "3212",
+		Rssi:  13,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+	sh.InsertSingleReading(mockReading)
+	sh.InsertSingleReading(mockReading2)
+
+	expectedMockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	serializedState, _ := sh.GetSerializedState()
+	actualState, _ := DeserializeState(serializedState)
+
+	sort.SliceStable(actualState.Readings, func(i, j int) bool {
+		return actualState.Readings[i].TagId < actualState.Readings[j].TagId
+	})
+
+	if !reflect.DeepEqual(actualState, expectedMockState) {
+		log.Printf("Insert\n: Expected %v\n, Got %v\n", actualState, expectedMockState)
+	}
+}
+
+func TestLessTimeDontInsertSingleReading(t *testing.T) {
+
+	time := timestamppb.Now().Seconds
+	time2 := time / 100
+
+	mockReading := &pb.Reading{
+		TagId: "MOCKTAG",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: time,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG",
+		RpId:  "321",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: time2,
+		},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+	sh.InsertSingleReading(mockReading)
+	sh.InsertSingleReading(mockReading2)
+	reading := sh.GetReading(mockReading.TagId)
+
+	assert.Equal(t, time, reading.Ts.Seconds)
+
+}
+
+func TestMoreTimeInsertSingleReading(t *testing.T) {
+
+	time := timestamppb.Now().Seconds
+	time2 := time * 100
+
+	mockReading := &pb.Reading{
+		TagId: "MOCKTAG",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: time,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG",
+		RpId:  "321",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: time2,
+		},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+
+	sh.InsertSingleReading(mockReading)
+	sh.InsertSingleReading(mockReading2)
+	reading := sh.GetReading(mockReading.TagId)
+	assert.Equal(t, time2, reading.Ts.Seconds)
+
+}
+
+func TestInsertMultipleReadings(t *testing.T) {
+
+	mockReading := &pb.Reading{
+		TagId: "MOCKTAG",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG2",
+		RpId:  "3212",
+		Rssi:  13,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading3 := &pb.Reading{
+		TagId: "MOCKTAG3",
+		RpId:  "3213",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading4 := &pb.Reading{
+		TagId: "MOCKTAG4",
+		RpId:  "3214",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	mockState2 := &pb.State{
+		TagId:    "669",
+		Readings: []*pb.Reading{mockReading3, mockReading4},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("669")
+	sh.InsertMultipleReadings(mockState)
+	sh.InsertMultipleReadings(mockState2)
+	expectedMockState := &pb.State{
+		TagId:    "669",
+		Readings: []*pb.Reading{mockReading, mockReading2, mockReading3, mockReading4},
+	}
+
+	serializedState, _ := sh.GetSerializedState()
+	actualState, _ := DeserializeState(serializedState)
+
+	sort.SliceStable(actualState.Readings, func(i, j int) bool {
+		return actualState.Readings[i].TagId < actualState.Readings[j].TagId
+	})
+
+	if !reflect.DeepEqual(actualState, expectedMockState) {
+		log.Printf("Insert\n: Expected %v\n, Got %v\n", actualState, expectedMockState)
+	}
+}
+
+func TestInsertMultipleReadingsOneState(t *testing.T) {
+
+	mockReading := &pb.Reading{
+		TagId: "666",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG2",
+		RpId:  "3212",
+		Rssi:  13,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+	sh.InsertMultipleReadings(mockState)
+	expectedMockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	serializedState, _ := sh.GetSerializedState()
+	actualState, _ := DeserializeState(serializedState)
+
+	sort.SliceStable(actualState.Readings, func(i, j int) bool {
+		return actualState.Readings[i].TagId < actualState.Readings[j].TagId
+	})
+
+	if !reflect.DeepEqual(actualState, expectedMockState) {
+		log.Printf("Insert\n: Expected %v\n, Got %v\n", actualState, expectedMockState)
+	}
+}
+
+func TestInsertMultipleReadingsMultipleStates(t *testing.T) {
+
+	mockReading := &pb.Reading{
+		TagId: "666",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG2",
+		RpId:  "3212",
+		Rssi:  13,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading3 := &pb.Reading{
+		TagId: "MOCKTAG3",
+		RpId:  "3213",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading4 := &pb.Reading{
+		TagId: "MOCKTAG4",
+		RpId:  "3214",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	mockState2 := &pb.State{
+		TagId:    "669",
+		Readings: []*pb.Reading{mockReading3, mockReading4},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+	sh.InsertMultipleReadings(mockState)
+	sh.InsertMultipleReadings(mockState2)
+	expectedMockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2, mockReading3, mockReading4},
+	}
+
+	serializedState, _ := sh.GetSerializedState()
+	actualState, _ := DeserializeState(serializedState)
+
+	sort.SliceStable(actualState.Readings, func(i, j int) bool {
+		return actualState.Readings[i].TagId < actualState.Readings[j].TagId
+	})
+
+	if !reflect.DeepEqual(actualState, expectedMockState) {
+		log.Printf("Insert\n: Expected %v\n, Got %v\n", actualState, expectedMockState)
+	}
+}
+
+func TestInsertMultipleReadingsSelfReadingLatest(t *testing.T) {
+
+	mockReading := &pb.Reading{
+		TagId: "666",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG2",
+		RpId:  "3212",
+		Rssi:  13,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading3 := &pb.Reading{
+		TagId: "669",
+		RpId:  "3213",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds / 10,
+		},
+	}
+
+	mockReading4 := &pb.Reading{
+		TagId: "MOCKTAG4",
+		RpId:  "3214",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	mockState2 := &pb.State{
+		TagId:    "669",
+		Readings: []*pb.Reading{mockReading3, mockReading4},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+	sh.InsertMultipleReadings(mockState)
+	sh.InsertMultipleReadings(mockState2)
+
+	fmt.Println(sh.readingsMap)
+	assert.Equal(t, int32(0), sh.readingsMap["666"].IsDirect)
+}
+
+func TestInsertMultipleReadingsSelfReadingNotLatest(t *testing.T) {
+
+	mockReading := &pb.Reading{
+		TagId: "666",
+		RpId:  "321",
+		Rssi:  69,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds / 10,
+		},
+	}
+
+	mockReading2 := &pb.Reading{
+		TagId: "MOCKTAG2",
+		RpId:  "3212",
+		Rssi:  13,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockReading3 := &pb.Reading{
+		TagId: "669",
+		RpId:  "3213",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds * 10,
+		},
+	}
+
+	mockReading4 := &pb.Reading{
+		TagId: "MOCKTAG4",
+		RpId:  "3214",
+		Rssi:  10,
+		Ts: &timestamp.Timestamp{
+			Seconds: timestamppb.Now().Seconds,
+		},
+	}
+
+	mockState := &pb.State{
+		TagId:    "666",
+		Readings: []*pb.Reading{mockReading, mockReading2},
+	}
+
+	mockState2 := &pb.State{
+		TagId:    "669",
+		Readings: []*pb.Reading{mockReading3, mockReading4},
+	}
+
+	sh := StateHandler{}
+	sh.InitStateHandler("666")
+	sh.InsertMultipleReadings(mockState)
+	sh.InsertMultipleReadings(mockState2)
+
+	fmt.Println(sh.readingsMap)
+	assert.Equal(t, int32(0), sh.readingsMap["669"].IsDirect)
 }
